requirements.txt
pandas
openpyxl
google-generativeai
python-dotenv
tqdm
streamlit
plotly
